:anchor: Added more UI for the spawnerManager; changed some inconsistencies

+++ Project Escargot/Game/Entity/LivingEntity/Player/Player.cpp
@@ -115,7 +115,7 @@ namespace Snail
 
 		if ((m_physicBodyRef->TriggeredMasks & MASK_ENEMY) == MASK_ENEMY)
 		{
-			int damages = STATE_GAME_BASE_ENEMIES_DAMAGES + STATE_GAME_BASE_ENEMIES_DAMAGES_PER_WAVE * (SpawnerManager::GetInstance()->GetWave() - 1);
+			float damages = (float) STATE_GAME_BASE_ENEMIES_DAMAGES + STATE_GAME_BASE_ENEMIES_DAMAGES_PER_WAVE * (SpawnerManager::GetInstance()->GetWave() - 1);
 			
 			m_TakeDamage(damages);
 		}
